+++ NEWS.md
@@ -1,7 +1,7 @@
 ### crunch 1.6.0
-* Check for new versions of the package on GitHub when the package loads.
-* Make a shallow `copy` of a variable. See `?copyVariable`. 
-* Correctly support `ds$var$subvar[filter] <- values`
+* Check for new version of the package on GitHub when the package is loaded.
+* Make a shallow `copy` of a variable. See `?copyVariable`.
+* Fix error in updating the values of a subvariable in an array.
 
 ### crunch 1.5.4
 * Gradually slow the polling of `/batches/` while waiting for an append to complete. Improves the performance of the append operation.